Move componentNameAsString to separate file
