+++ src/scripts/react/common/ComponentName.jsx
@@ -4,13 +4,6 @@ const shouldShowType = (component) => {
   return component.get('type') === 'extractor' || component.get('type') === 'writer';
 };
 
-export const componentNameAsString = (component, options = {showType: false}) => {
-  if (options && options.showType && shouldShowType(component)) {
-    return component.get('name') + ' ' + component.get('type');
-  }
-  return component.get('name');
-};
-
 export default React.createClass({
   propTypes: {
     component: PropTypes.object.isRequired,
